requirements.txt
matplotlib==3.7.2
numpy==1.25.2
pandas==2.1.0
psycopg2-binary==2.9.7
seaborn==0.12.2
scikit-learn==1.3.0
SQLAlchemy==2.0.20
xgboost==2.0.0
tqdm==4.66.1
requests==2.31.0
pvlib==0.10.2

